requirements.txt
Django==5.1.1
djangorestframework==3.15.2
odmantic==1.0.2
pymongo==4.6.1
web3==6.0.0
python-dotenv
